requirements.txt
openai>=1.30.0
playwright>=1.44.0
supabase>=2.4.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.29.0
httpx>=0.27.0

# v2
